+++ config/medium/Criminal/code/CGCode.js
@@ -1,0 +1,10 @@
+const H = parseInt(readline());
+const W = parseInt(readline());
+for (let i = 0; i < H; i++) {
+    const row = readline();
+}
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('answer');