+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - configuration for "🏐 🏖 Beach Volleyball 🏐 🏖".
 - configuration for "Rocket engines".
 - configuration for "ASCII cube".
+- configuration for "Criminal".
 
 ## [1.9.0] - 2024-11-30
 ### Added